restaurant_booking_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-dotenv
alembic
pytest
httpx
pytest-asyncio
